rename variables in kms
